# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OpenAPI",
    "version": "1.0.0",
    "description": "BaseApp Next.js Fullstack Template"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "GetAccountCredentialsQueryParams": {
        "type": "object",
        "properties": {
          "vaultId": {
            "type": "string",
            "description": "Vault ID"
          },
          "search": {
            "type": "string",
            "description": "Search",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "account",
              "bankAccount"
            ],
            "nullable": true
          }
        },
        "required": [
          "vaultId"
        ]
      },
      "GetAccountCredentialsResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/GetAccountCredentialsResponseItem"
        }
      },
      "CreateAccountCredentialsBody": {
        "type": "object",
        "properties": {
          "vaultId": {
            "type": "string",
            "description": "Vault ID"
          },
          "name": {
            "type": "string",
            "description": "Name"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "nullable": true
          },
          "username": {
            "type": "string",
            "description": "Username",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "Password",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "description": "Notes",
            "nullable": true
          },
          "bankAccount": {
            "type": "object",
            "properties": {
              "bankName": {
                "type": "string",
                "description": "Bank Name"
              },
              "owners": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Owners"
              },
              "beneficiaries": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Beneficiaries"
              },
              "accountNumber": {
                "type": "string",
                "description": "Account Number",
                "nullable": true
              },
              "aba": {
                "type": "string",
                "description": "ABA",
                "nullable": true
              },
              "swift": {
                "type": "string",
                "description": "SWIFT",
                "nullable": true
              },
              "bankAddress": {
                "type": "string",
                "description": "Bank Address",
                "nullable": true
              },
              "beneficiaryAddress": {
                "type": "string",
                "description": "Beneficiary Address",
                "nullable": true
              }
            },
            "required": [
              "bankName",
              "owners",
              "beneficiaries"
            ],
            "nullable": true
          }
        },
        "required": [
          "vaultId",
          "name"
        ]
      },
      "GetUserPathParams": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          }
        },
        "required": [
          "id"
        ]
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "username": {
                "type": "string",
                "nullable": true
              },
              "name": {
                "type": "string"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": [
              "id",
              "email",
              "name",
              "createdAt",
              "updatedAt"
            ]
          }
        },
        "required": [
          "user"
        ]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "username": {
                "type": "string",
                "nullable": true
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "email",
              "name"
            ]
          }
        },
        "required": [
          "user"
        ]
      },
      "GetUsersReponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "username": {
              "type": "string",
              "nullable": true
            },
            "name": {
              "type": "string"
            }
          },
          "required": [
            "id",
            "email",
            "name"
          ]
        }
      },
      "GetVaultPathParams": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Vault ID"
          }
        },
        "required": [
          "id"
        ]
      },
      "GetVaultResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "owner": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "username": {
                "type": "string",
                "nullable": true
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "email",
              "name"
            ]
          },
          "members": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "username": {
                  "type": "string",
                  "nullable": true
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "email",
                "name"
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "owner",
          "members"
        ]
      },
      "JoinVaultBody": {
        "type": "object",
        "properties": {
          "vaultKey": {
            "type": "string",
            "description": "Vault Key"
          }
        },
        "required": [
          "vaultKey"
        ]
      },
      "JoinVaultResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "RemoveMemberFromVaultBody": {
        "type": "object",
        "properties": {
          "vaultId": {
            "type": "string",
            "description": "Vault ID"
          },
          "userId": {
            "type": "string",
            "description": "User ID"
          }
        },
        "required": [
          "vaultId",
          "userId"
        ]
      },
      "GetVaultsResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/GetVaultsResponseItem"
        }
      },
      "CreateVaultBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "name"
        ]
      },
      "GetVaultsResponseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "ownerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "ShareVaultBody": {
        "type": "object",
        "properties": {
          "vaultId": {
            "type": "string",
            "description": "Vault ID"
          }
        },
        "required": [
          "vaultId"
        ]
      },
      "ShareVaultResponse": {
        "type": "object",
        "properties": {
          "vaultKey": {
            "type": "string",
            "description": "Vault Key"
          }
        },
        "required": [
          "vaultKey"
        ]
      },
      "GetAccountCredentialsResponseItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "bankAccount": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      }
    },
    "responses": {}
  },
  "outputDir": "./public",
  "paths": {
    "/account-credentials": {
      "get": {
        "operationId": "get-account-credentials",
        "summary": "Get all account credentials",
        "description": "Fetches all account credentials for the current user",
        "tags": [
          "Account-credentials"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "vaultId",
            "schema": {
              "type": "string",
              "description": "Vault ID"
            },
            "required": false,
            "description": "Vault ID"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string",
              "description": "Search"
            },
            "required": false,
            "description": "Search"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "account",
                "bankAccount"
              ]
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAccountCredentialsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-account-credentials",
        "summary": "Create a new account credentials",
        "description": "Creates a new account credentials for the current user",
        "tags": [
          "Account-credentials"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountCredentialsBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/studio": {
      "get": {
        "operationId": "get-studio",
        "summary": "Internal route - not for documentation",
        "description": "",
        "tags": [
          "Studio"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-studio",
        "summary": "Internal route - not for documentation",
        "description": "",
        "tags": [
          "Studio"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "operationId": "get-users",
        "summary": "Get all users",
        "description": "Fetches all users",
        "tags": [
          "Users"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUsersReponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "get-users-{id}",
        "summary": "Get user by ID",
        "description": "Fetches user information by ID",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "User ID"
            },
            "required": true,
            "description": "User ID",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "operationId": "get-users-me",
        "summary": "Get current user information",
        "description": "Fetches current user information",
        "tags": [
          "Users"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/vaults": {
      "get": {
        "operationId": "get-vaults",
        "summary": "Get all vaults",
        "description": "Fetches all vaults for the current user",
        "tags": [
          "Vaults"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVaultsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-vaults",
        "summary": "Create a new vault",
        "description": "Creates a new vault for the current user",
        "tags": [
          "Vaults"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVaultBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVaultsResponseItem"
                }
              }
            }
          }
        }
      }
    },
    "/vaults/{id}": {
      "get": {
        "operationId": "get-vaults-{id}",
        "summary": "Get vault by ID",
        "description": "Fetches vault information by ID",
        "tags": [
          "Vaults"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Vault ID"
            },
            "required": true,
            "description": "Vault ID",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVaultResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-vaults-{id}",
        "summary": "Delete a vault",
        "description": "Deletes a vault by ID",
        "tags": [
          "Vaults"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Vault ID"
            },
            "required": true,
            "description": "Vault ID",
            "example": "123"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/vaults/join": {
      "post": {
        "operationId": "post-vaults-join",
        "summary": "Join a vault",
        "description": "Joins a vault by vault key",
        "tags": [
          "Vaults"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinVaultBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinVaultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/vaults/remove-member": {
      "post": {
        "operationId": "post-vaults-remove-member",
        "summary": "Remove a member from a vault",
        "description": "Removes a member from a vault by vault ID and user ID",
        "tags": [
          "Vaults"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveMemberFromVaultBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/vaults/share": {
      "post": {
        "operationId": "post-vaults-share",
        "summary": "Share a vault",
        "description": "Shares a vault by creating an invitation",
        "tags": [
          "Vaults"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareVaultBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareVaultResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}